fix: Split WeierstrassDecompress into two chips (#138)
* fix: Split WeierstrassDecompress into two chips

This commit deletes the previously-generic WeierstrassDecompressChip and
replaces it with two non-generic variants, one for Secp256k1 and one for
BLS12-381.

The reasoning behind this is described in #130, but in short:

- WeierstrassDecompressChip was written for Secp256k1, and it assumes a
certain meaning of parity for the y selection bit (is_odd column)
- This is not the same meaning used for the y bit in BLS12-381, which
uses a bit to indicate if the y coordinate is the lexicographically
largest of the two candidates
- This mismatch was annoying to combine into a single fully generic
chip, and the benefits of the WeierstrassDecompress chip being generic
are limited since it also assumes that `a = 0`

Besides removing the generic code, this commit also makes the BLS12-381
G1 decompress chip receive the full compressed bits from the syscall,
including compression flags. The flags are now handled inside the AIR
constraints. The handling of the point at infinity is still done
out-of-circuit.

To check whether `y > (p-1)/2`, we instead calculate the value of `2*y`
and check its parity by looking at its LSBit. If `2*y` is even, then `y
<= (p-1)/2` and if `2*y` is odd, then `y > (p-1)/2` due to the behavior
of modular reduction. This check is cheaper to do in-circuit than a full
bit decomposition.

Additional care is taken to ensure that the MSByte of `x` is overwritten
to clear any compression flags.

* fix: Remove the unmasking of flags in G1 non-zkvm

With the new behavior, the output of the `syscall_g1_decompress` should
match the `to_uncompressed` output exactly.

* chore: Address review nits and more docstrings

* fix: Fix and recompile tests/bls12381-mul