refactor: Make Chips responsible for what they need from Records
Currently the `MachineAir` trait requires every chip to define how it interacts with an `ExecutionRecord` associated type.

This forces any chip that implements `MachineAir` to only interact with one specific implementation of an `ExecutionRecord`
(as fixed at the moment of choosing that associated type).

We would like those chips to be reused in more varied ways, and the following starts the changes towards accomplishing that.

We set up the general framework with:
```rust
/// A description of the events related to this AIR.
pub trait WithEvents<'a>: Sized {
    /// output of a functional lens from the Record to
    /// refs of those events relative to the AIR.
    type Events: 'a;
}

pub trait EventLens<T: for<'a> WithEvents<'a>> {
    fn events(&self) -> <T as WithEvents>::Events;
}

pub trait MachineAir<F: Field>: BaseAir<F> + for<'a> WithEvents<'a> {
    ...
     fn generate_trace<EL: EventLens<Self>>(&self, input: &EL, output: &mut ExecutionRecord) -> RowMajorMatrix<F>;
    ...
}
```
(the change to output is similar and pending)

then in `AddSubChip`:
```rust
impl<'a> WithEvents<'a> for AddSubChip {
    type Events = (
        // add events
        &'a [AluEvent],
        // sub events
        &'a [AluEvent],
    );
}
```

In the `ExecutionRecord`:
```rust
impl EventLens<AddSubChip> for ExecutionRecord {
    fn events(&self) -> <AddSubChip as crate::air::WithEvents>::Events {
        (&self.add_events, &self.sub_events)
    }
}
```

In `generate_trace`:
```rust
    fn generate_trace<EL: EventLens<Self>>(
        &self,
        input: &EL,
        output: &mut EL,
    ) -> RowMajorMatrix<F> {
        let (add_events, sub_events) = input.events();
        // Generate the rows for the trace.
        let chunk_size = std::cmp::max(
            (add_events.len() + sub_events.len()) / num_cpus::get(),
            1,
        );
        let merged_events = add_events
            .iter()
            .chain(sub_events.iter())
            .collect::<Vec<_>>();
    ...
```