fix: Revert some memory chip constraints
The MemoryProgram chip was modified to only be included in the first
shard. However, the rust out-of-circuit verifier was not updated to
account for this since it still expects the chip to be present in every
shard, and the in-circuit recursive verifier was updated with
constraints that are only valid if the core proof has a single shard.

This commit temporarily comments out these additional checks, but this
should be reverted when the verifier and Memory chip-related fixes are
integrated. This is related to the batch of PRs for issue #38

It might be necessary to revert commit
6aea5dcbe4290bdabbfcaded3f03288329a27d73
after the proper fixes are incorporated.