fix: Reorganize `AbsorbWorkspace` columns
The two columns in `AbsorbWorkspace` introduced in
`210423a5403b25319da2b60dce78e0d9d107d0aa` seemed to expose some
underlying issue which made the gnark circuit fail to accept a proof
even though the execution did not trigger any `TRAP`s. It seems like the
placement of these two columns in this struct were the reason why it was
failing, possibly some unfortunate interaction of `AlignedBorrow`, Rust
unions, or something else.

The issue exposed itself as a gnark constraint failure, and the specific
constraint that was failing was the one responsible for verifying the
constraints for the `Multi` chip. The `Multi` chip can have either a
`FriFoldChip` or a `Poseidon2WideChip` for each row. This might help
explain why this change was leading to constraint failures.